pyproject.toml
[project]
name = "edugraph-qwen3-finetune"
version = "0.1.0"
description = "Fine-tuning Qwen3-VL for educational material classification."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    # PyTorch dependencies for CUDA 12.6
    "torch~=2.6.0",
    "torchvision~=0.21.0",

    # Transformers and fine-tuning ecosystem
    "transformers @ git+https://github.com/huggingface/transformers",
    "deepspeed~=0.17.1",
    "flash-attn~=2.7.4.post1",
    #"triton~=3.2.0",
    "accelerate~=1.7.0",
    "torchcodec~=0.2",
    "peft~=0.17.1",
    "trl~=0.7.11",

    # Core ML/data libraries
    "datasets~=2.18.0",
    "evaluate~=0.4.2",
    "einops~=0.7.0",
    "scikit-learn~=1.4.0",
    "pandas~=2.2.0",
    "numpy~=1.26.0",

    # Utilities
    "Pillow~=10.2.0",
    "opencv-python~=4.9.0.80",
    "qwen-vl-utils==0.0.14",
    "python-dotenv~=1.0.0",
    "rdflib~=7.0.0",
    "pdf2image~=1.17.0",
]

# Define a named index for the PyTorch CUDA packages
[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

# Map specific packages to the named index
[tool.uv.sources]
torch = [
  { index = "pytorch-cuda" },
]
torchvision = [
  { index = "pytorch-cuda" },
]
